requirements.txt
Django==2.1.7
django-environ==0.4.5
pytz==2018.9
uWSGI==2.0.18
